created new CoursItem composable, extracted and refactored sub-composables
