+++ core/ui/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/ui/material/colors/CourseColors.kt
@@ -1,0 +1,13 @@
+package de.tum.informatics.www1.artemis.native_app.core.ui.material.colors
+
+import androidx.compose.ui.graphics.Color
+
+object CourseColors {
+
+    object CircularCourseProgress {
+        val circle: Color
+            get() = Color(0xffdc3545)
+        val track: Color
+            get() = Color(0xff28a745)
+    }
+}
