+++ feature/dashboard/src/main/res/values/course_overview_ui_strings.xml
@@ -2,11 +2,6 @@
 <resources>
     <string name="course_overview_title">Courses</string>
 
-    <!--  The placeholders are digits, but are supplied as strings, as the number formatting is handled in code.  -->
-    <string name="course_overview_course_progress">%1$sP/%2$sP (%3$s)</string>
-    <string name="course_overview_course_progress_pts">%1$s /%2$s pts</string>
-    <string name="course_overview_course_progress_percentage">%1$s</string>
-
     <string name="course_overview_action_register">Register for course</string>
     <string name="courses_loading_loading">Loading dashboard…</string>
     <string name="courses_loading_suspended">Your internet connection may be unstable. Could not load dashboard.</string>
